Use size_t instead of int to avoid mixing signedness

+++ mo-machine-outlier/mo.cpp
@@ -802,7 +802,7 @@ partition_single_side(vector<TupleWrapper<T>> &tuple_wrapper_list, size_t left,
 template<typename T>
 static inline TupleWrapper<T>
 bfprt_wrapper(vector<TupleWrapper<T>> &tuple_wrapper_list, size_t i,
-              size_t left, int right) {
+              size_t left, size_t right) {
     assert(left <= right);
     assert(left <= i);
     assert(i <= right);
